src/main/java/com/tw/vapsi/biblioteca/service/BookService.java
package com.tw.vapsi.biblioteca.service;

import com.tw.vapsi.biblioteca.model.Book;
import com.tw.vapsi.biblioteca.repository.BookRepository;
import lombok.AllArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Collections;
import java.util.List;

@Service
@AllArgsConstructor
public class BookService {


    @Autowired
    private BookRepository bookRepository;


    public List<Book> fetchAllBooks() {
        return (List<Book>) bookRepository.findAll();
    }
}
